Implement and export PLAwait
A PLAwait value tells Dart code evaluating a PiLisp program that it
should `await` the result.

This falls short of a generally-useful await functionality within
PiLisp code, since it does not allow a PiLisp programmer to await a
value within a PiLisp program and then continue execution.

PiLisp would need to support a form of continuations to continue
having synchronous eval and support ad hoc await within PiLisp
programs. This is outside the scope of the language implementation at
this time.

The pilisp-native extension to pilisp has a `future/wait-for` function
that supports synchronously waiting on Future values within a PiLisp
program using the (now deprecated) `dart:cli.waitFor` method.