Fix PiLisp stack traces
At the most fundamental level, the implementation before this commit
was popping stack frames in finally blocks, preventing them from being
surfaced at the top of the call stack when printing out error
messages.

This commit introduces a new PLInvocationException class, which stores
the frames when called.

The sites where invocable objects are invoked have been updated with
try/catch logic that builds instance of PLInvocationException, saving
the current stack frames and avoiding re-catching
PLInvocationException exceptions (so that the innermost exception,
with its full stack trace, bubbles up as expected).

There is an existing PLException object which also saves stack
information. This class needs to be reconsidered in light of how
PLInvocationException is structured, but for now I do not plan to
consolidate these two, as PLException is the return value produced by
ex-info and a separate category from exceptions thrown by underlying
Dart implementation.